First Servlet
I created a servlet as a test. Copy the wtpwebapps into your
installation of Tomcat. Start Tomcat. You should be able to view the
application at localhost:8080/FirstServlet/HelloWorldServles. The name
should be HelloWorldServlet but I misnamed the project when I created
the project in Eclipse. It ought to run anyway.