FeedService.java update
FeedService.java now has a method called GetMethod. GetMethod takes two
url parameters: terrorParam1 and terrorParam2. You pass the values by
using the URL
http://localhost:8080/TerrorismData/TerrorData/WebService/GetMethod?terr
orParam1=this&terrorParam2=that. Replace this and that with whatever
string value you want to pass to the function. Once the values are
passed into the function, the function returns as a plain text
document: These are your two params---->this and that. (Assuming that
this and that are the two values contained in the URL)