Created method that pulls years and attacks for a given group
The Project.java, ProjectManager.java and FeedService.java files have
been updated. There is a new method called GetAttacksByYear in
Project.java. The method returns an ArrayList of Maps with the key
being the year and the value being the number of attacks for that year.
FeedService.java uses the GSON library to return that result as a JSON
string.